backend/lambdas/loader/requirements.txt
psycopg2-binary
boto3
smart-open
opensearch-py
requests
requests-aws4auth
